fix: address PR audit issues and update tina versions
- Fix AccordionBlock grid column bug (item.image → imageData)
- Deduplicate ImageMetadata interface into shared export
- Fix accordion mapping type from string | ImageMetadata to ImageMetadata
- Consolidate TinaCloud URL normalization into resolve-image-src utility
- Restore img fallback handler delegating to ImageEmbed
- Update tinacms 3.7.3 → 3.8.1, @tinacms/cli 2.2.3 → 2.3.1
