+++ src/utils/resolve-image-src.ts
@@ -1,3 +1,14 @@
+const TINA_CLOUD_ASSET_REGEX = /^https?:\/\/assets\.tina\.io\/[^/]+\/(.+)$/;
+
+/** Extract local path from TinaCloud CDN URLs, or return the path as-is. */
+export function normalizeTinaCloudSrc(src: string): string {
+  const match = src.match(TINA_CLOUD_ASSET_REGEX);
+  if (match) {
+    return `/${match[1]}`;
+  }
+  return src;
+}
+
 /** Prepend NEXT_PUBLIC_BASE_PATH for local image paths; pass through external URLs. */
 export function resolveImageSrc(src: string): string {
   if (
