+++ tina/collections/image-metadata.tsx
@@ -1,10 +1,10 @@
-/**
- * Reusable TinaCMS schema fields for images with metadata.
- * Captures width, height, and alt text alongside the image path.
- *
- * Used by accordion and scroll-showcase templates alongside
- * the ImageWithMetadataField custom component.
- */
+export interface ImageMetadata {
+  src: string;
+  width?: number;
+  height?: number;
+  alt?: string;
+}
+
 export const ImageWithMetadataFields = [
   {
     name: "src",
